# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: DaYu Platform HTTP API
  version: 1.0.0-draft
  description: |
    大禹平台第一版 HTTP 契约。所有业务时间均为 UTC ISO-8601。
    浏览器先通过 GET /api/v1/session 获取与当前 Session 绑定的 CSRF Token；
    所有 POST、PUT、PATCH、DELETE 请求都必须携带 X-CSRF-TOKEN。
    本契约不暴露服务器绝对路径、NC 相对路径或 Nginx 内部下载位置。
servers:
  - url: /
...
paths:
...
  /api/v1/session:
    get:
      tags: [Identity]
      summary: 获取当前 Session、身份和 CSRF Token
      security: []
      responses:
        '200':
          description: 未登录时也返回 200，并建立匿名 Session
          headers:
            Cache-Control:
              schema: {type: string, const: no-store}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SessionView'}
    post:
      tags: [Identity]
      summary: 登录并建立认证 Session
      security:
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: 登录成功，响应携带当前认证 Session 对应的 CSRF Token
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SessionView'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
        '429': {$ref: '#/components/responses/RateLimited'}
    delete:
      tags: [Identity]
      summary: 退出并销毁当前 Session
      security:
        - sessionCookie: []
          csrfToken: []
      responses:
        '204': {description: Session 和原 CSRF Token 已失效}
        '401': {$ref: '#/components/responses/Unauthenticated'}
  /api/v1/users:
    post:
      tags: [Identity]
      summary: 注册普通用户并在事务提交后建立认证 Session
      security:
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterRequest'}
      responses:
        '201':
          description: 用户落库成功且认证 Session 已建立
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SessionView'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/downloads:
    post:
      tags: [Download]
      summary: 登记用途并申请短期 NC 下载授权
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DownloadRequest'}
      responses:
        '201':
          description: 授权成功；downloadUrl 不包含服务器路径
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DownloadAuthorization'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '410': {$ref: '#/components/responses/AssetGone'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/downloads/{downloadEventId}/content:
    get:
      tags: [Download]
      summary: 使用短期授权下载 NC 文件
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/DownloadEventIdPath'
      responses:
        '200':
          description: Nginx 通过内部转发传输文件
          headers:
            Content-Disposition: {schema: {type: string}}
            Content-Length: {schema: {type: integer, format: int64}}
            Cache-Control: {schema: {type: string, const: 'private, no-store'}}
          content:
            application/x-netcdf:
              schema: {type: string, format: binary}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '410': {$ref: '#/components/responses/AssetGone'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: DAYUSESSID
      description: HttpOnly、SameSite=Lax；生产 HTTPS 下 Secure=true
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: 必须与当前匿名或认证 Session 匹配
  parameters:
...
    DownloadEventIdPath:
      name: downloadEventId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
  responses:
    Unauthenticated:
      description: 未登录或认证已失效
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Forbidden:
      description: 无权限、CSRF 失败或路径安全校验失败
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    NotFound:
      description: 产品、资产或事件不存在；公开接口也用它隐藏未发布产品
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    AssetGone:
      description: 已索引资产对应文件丢失或短期授权过期
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Conflict:
      description: 唯一性、状态机或并发任务冲突
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    ValidationFailed:
      description: 参数格式正确但不满足业务校验
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    RateLimited:
      description: 请求频率超限
      headers:
        Retry-After: {schema: {type: integer, minimum: 1}}
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
  schemas:
...
    UtcInstant: {type: string, format: date-time, pattern: 'Z$'}
...
    UserRole: {type: string, enum: [USER, ADMIN]}
...
    ErrorResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code: {type: string}
        message: {type: string}
        traceId: {type: string}
        details: {type: object, additionalProperties: true}
...
    AuthenticatedUser:
      type: object
      required: [userId, email, organization, role]
      properties:
        userId: {type: integer, format: int64}
        email: {type: string, format: email}
        organization: {type: string}
        role: {$ref: '#/components/schemas/UserRole'}
    SessionView:
      type: object
      required: [authenticated, user, csrfToken]
      properties:
        authenticated: {type: boolean}
        user: {oneOf: [{$ref: '#/components/schemas/AuthenticatedUser'}, {type: 'null'}]}
        csrfToken: {type: string, minLength: 1}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email, maxLength: 254}
        password: {type: string, minLength: 1, maxLength: 128, writeOnly: true}
    RegisterRequest:
      type: object
      required: [email, password, organization]
      properties:
        email: {type: string, format: email, maxLength: 254}
        password: {type: string, minLength: 10, maxLength: 128, writeOnly: true}
        organization: {type: string, minLength: 2, maxLength: 255}
    DownloadRequest:
      type: object
      required: [assetId, purpose]
      properties:
        assetId: {type: integer, format: int64, minimum: 1}
        purpose: {type: string, minLength: 10, maxLength: 2000}
    DownloadAuthorization:
      type: object
      required: [downloadEventId, downloadUrl, fileName, expectedBytes, expiresAt]
      properties:
        downloadEventId: {type: integer, format: int64}
        downloadUrl: {type: string, format: uri-reference}
        fileName: {type: string}
        expectedBytes: {type: integer, format: int64, minimum: 0}
        expiresAt: {$ref: '#/components/schemas/UtcInstant'}
...